+++ src/main/java/com/jsorrell/carpetskyadditions/fakes/CamelInterface.java
@@ -1,0 +1,7 @@
+package com.jsorrell.carpetskyadditions.fakes;
+
+public interface CamelInterface {
+    void carpetSkyAdditions$makeTraderCamel();
+
+    void carpetSkyAdditions$makeStandaloneCamel();
+}