+++ src/main/java/com/jsorrell/carpetskyadditions/mixin/WanderingTraderAccessor.java
@@ -1,0 +1,12 @@
+package com.jsorrell.carpetskyadditions.mixin;
+
+import net.minecraft.core.BlockPos;
+import net.minecraft.world.entity.npc.WanderingTrader;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Accessor;
+
+@Mixin(WanderingTrader.class)
+public interface WanderingTraderAccessor {
+    @Accessor
+    BlockPos getWanderTarget();
+}